Move RocksDbTransaction into a separate class
