+++ storage/src/main/java/tech/pegasys/teku/storage/server/rocksdb/RocksDbInstance.java
@@ -21,7 +21,6 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.locks.ReentrantLock;
 import java.util.function.Consumer;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
@@ -31,7 +30,6 @@ import org.rocksdb.ColumnFamilyHandle;
 import org.rocksdb.RocksDBException;
 import org.rocksdb.RocksIterator;
 import org.rocksdb.TransactionDB;
-import org.rocksdb.WriteOptions;
 import tech.pegasys.teku.storage.server.ShuttingDownException;
 import tech.pegasys.teku.storage.server.kvstore.ColumnEntry;
 import tech.pegasys.teku.storage.server.kvstore.KvStoreAccessor;
@@ -44,7 +42,7 @@ public class RocksDbInstance implements KvStoreAccessor {
   private final ColumnFamilyHandle defaultHandle;
   private final ImmutableMap<KvStoreColumn<?, ?>, ColumnFamilyHandle> columnHandles;
   private final List<AutoCloseable> resources;
-  private final Set<Transaction> openTransactions = new HashSet<>();
+  private final Set<RocksDbTransaction> openTransactions = new HashSet<>();
 
   private final AtomicBoolean closed = new AtomicBoolean(false);
 
@@ -144,7 +142,8 @@ public class RocksDbInstance implements KvStoreAccessor {
   @MustBeClosed
   public synchronized KvStoreTransaction startTransaction() {
     assertOpen();
-    Transaction tx = new Transaction(db, defaultHandle, columnHandles, openTransactions::remove);
+    RocksDbTransaction tx =
+        new RocksDbTransaction(db, defaultHandle, columnHandles, openTransactions::remove);
     openTransactions.add(tx);
     return tx;
   }
@@ -170,7 +169,7 @@ public class RocksDbInstance implements KvStoreAccessor {
   @Override
   public synchronized void close() throws Exception {
     if (closed.compareAndSet(false, true)) {
-      for (Transaction openTransaction : openTransactions) {
+      for (RocksDbTransaction openTransaction : openTransactions) {
         openTransaction.closeViaDatabase();
       }
       db.syncWal();
@@ -183,168 +182,6 @@ public class RocksDbInstance implements KvStoreAccessor {
   private void assertOpen() {
     if (closed.get()) {
       throw new ShuttingDownException();
-    }
-  }
-
-  public static class Transaction implements KvStoreTransaction {
-    private final ColumnFamilyHandle defaultHandle;
-    private final ImmutableMap<KvStoreColumn<?, ?>, ColumnFamilyHandle> columnHandles;
-    private final org.rocksdb.Transaction rocksDbTx;
-    private final WriteOptions writeOptions;
-
-    private final ReentrantLock lock = new ReentrantLock();
-    private final AtomicBoolean closedViaDatabase = new AtomicBoolean(false);
-    private final Consumer<Transaction> onClosed;
-    private boolean closed = false;
-
-    private Transaction(
-        final TransactionDB db,
-        final ColumnFamilyHandle defaultHandle,
-        final ImmutableMap<KvStoreColumn<?, ?>, ColumnFamilyHandle> columnHandles,
-        final Consumer<Transaction> onClosed) {
-      this.defaultHandle = defaultHandle;
-      this.columnHandles = columnHandles;
-      this.writeOptions = new WriteOptions();
-      this.rocksDbTx = db.beginTransaction(writeOptions);
-      this.onClosed = onClosed;
-    }
-
-    @Override
-    public <T> void put(KvStoreVariable<T> variable, T value) {
-      applyUpdate(
-          () -> {
-            final byte[] serialized = variable.getSerializer().serialize(value);
-            try {
-              rocksDbTx.put(defaultHandle, variable.getId().toArrayUnsafe(), serialized);
-            } catch (RocksDBException e) {
-              throw RocksDbExceptionUtil.wrapException("Failed to put variable", e);
-            }
-          });
-    }
-
-    @Override
-    public <K, V> void put(KvStoreColumn<K, V> column, K key, V value) {
-      applyUpdate(
-          () -> {
-            final byte[] keyBytes = column.getKeySerializer().serialize(key);
-            final byte[] valueBytes = column.getValueSerializer().serialize(value);
-            final ColumnFamilyHandle handle = columnHandles.get(column);
-            try {
-              rocksDbTx.put(handle, keyBytes, valueBytes);
-            } catch (RocksDBException e) {
-              throw RocksDbExceptionUtil.wrapException("Failed to put column data", e);
-            }
-          });
-    }
-
-    @Override
-    public <K, V> void put(KvStoreColumn<K, V> column, Map<K, V> data) {
-      applyUpdate(
-          () -> {
-            final ColumnFamilyHandle handle = columnHandles.get(column);
-            for (Map.Entry<K, V> kvEntry : data.entrySet()) {
-              final byte[] key = column.getKeySerializer().serialize(kvEntry.getKey());
-              final byte[] value = column.getValueSerializer().serialize(kvEntry.getValue());
-              try {
-                rocksDbTx.put(handle, key, value);
-              } catch (RocksDBException e) {
-                throw RocksDbExceptionUtil.wrapException("Failed to put column data", e);
-              }
-            }
-          });
-    }
-
-    @Override
-    public <K, V> void delete(KvStoreColumn<K, V> column, K key) {
-      applyUpdate(
-          () -> {
-            final ColumnFamilyHandle handle = columnHandles.get(column);
-            try {
-              rocksDbTx.delete(handle, column.getKeySerializer().serialize(key));
-            } catch (RocksDBException e) {
-              throw RocksDbExceptionUtil.wrapException("Failed to delete key", e);
-            }
-          });
-    }
-
-    @Override
-    public <T> void delete(KvStoreVariable<T> variable) {
-      applyUpdate(
-          () -> {
-            try {
-              rocksDbTx.delete(defaultHandle, variable.getId().toArrayUnsafe());
-            } catch (RocksDBException e) {
-              throw RocksDbExceptionUtil.wrapException("Failed to delete variable", e);
-            }
-          });
-    }
-
-    @Override
-    public void commit() {
-      applyUpdate(
-          () -> {
-            try {
-              this.rocksDbTx.commit();
-            } catch (RocksDBException e) {
-              throw RocksDbExceptionUtil.wrapException("Failed to commit transaction", e);
-            } finally {
-              close();
-            }
-          });
-    }
-
-    @Override
-    public void rollback() {
-      applyUpdate(
-          () -> {
-            try {
-              this.rocksDbTx.commit();
-            } catch (RocksDBException e) {
-              throw RocksDbExceptionUtil.wrapException("Failed to commit transaction", e);
-            } finally {
-              close();
-            }
-          });
-    }
-
-    private void applyUpdate(final Runnable operation) {
-      lock.lock();
-      try {
-        assertOpen();
-        operation.run();
-      } finally {
-        lock.unlock();
-      }
-    }
-
-    private void assertOpen() {
-      if (closed) {
-        if (closedViaDatabase.get()) {
-          throw new ShuttingDownException();
-        } else {
-          throw new IllegalStateException("Attempt to update a closed transaction");
-        }
-      }
-    }
-
-    private void closeViaDatabase() {
-      closedViaDatabase.set(true);
-      close();
-    }
-
-    @Override
-    public void close() {
-      lock.lock();
-      try {
-        if (!closed) {
-          closed = true;
-          onClosed.accept(this);
-          writeOptions.close();
-          rocksDbTx.close();
-        }
-      } finally {
-        lock.unlock();
-      }
     }
   }
 }
